almost done

+++ utils.cpp
@@ -195,7 +195,7 @@ String check_display_value(double value, double undef, uint8_t len, uint8_t str_
 //String check_display_value(double value, double undef, char* len, uint8_t str_len) {
 
 	RESERVE_STRING(s, 15);
-	s = (value != undef ? String(value, len) : String("-"));
+	s = (value != undef ? String(value, (int)len) : String("-"));
 	//s = (value != undef ? String(value, (int)len) : String("-"));
 
 	while (s.length() < str_len) {
